+++ unittests/model_tests/application_graph_tests/test_application_vertex.py
@@ -5,7 +5,7 @@ from pacman.model.graphs.common import Slice
 from pacman.model.graphs.machine import SimpleMachineVertex
 
 # unit tests imports
-from uinit_test_objects.test_vertex import TestVertex
+from uinit_test_objects.simple_test_vertex import TestVertex
 
 # general imports
 import unittest
