+++ unittests/model_tests/graph_mapper_tests/test_graph_mapper.py
@@ -3,8 +3,8 @@ tests for graph mapper
 """
 
 # unit test objects
-from uinit_test_objects.test_edge import TestEdge
-from uinit_test_objects.test_vertex import TestVertex
+from uinit_test_objects.simple_test_edge import TestEdge
+from uinit_test_objects.simple_test_vertex import TestVertex
 
 # pacman imports
 from pacman.model.graphs.common import Slice, GraphMapper
